Rename `PermissionManager`
