+++ src/MediaWiki/PermissionManager.php
@@ -2,221 +2,69 @@
 
 namespace SMW\MediaWiki;
 
-use SMW\DataValues\AllowsPatternValue;
-use SMW\Protection\ProtectionValidator;
 use Title;
 use User;
 
 /**
  * @license GNU GPL v2+
- * @since 2.4
+ * @since 3.2
  *
  * @author mwjames
  */
 class PermissionManager {
 
 	/**
-	 * @var ProtectionValidator
+	 * @var PermissionManager|null
 	 */
-	private $protectionValidator;
+	private $permissionManager;
 
 	/**
-	 * @var PermissionExaminer
-	 */
-	private $permissionExaminer;
-
-	/**
-	 * @var []
-	 */
-	private $errors = [];
-
-	/**
-	 * @since 2.5
+	 * @since 3.2
 	 *
-	 * @param ProtectionValidator $protectionValidator
-	 * @param PermissionExaminer $permissionExaminer
+	 * @param PermissionManager|null $permissionManager
 	 */
-	public function __construct( ProtectionValidator $protectionValidator, PermissionExaminer $permissionExaminer ) {
-		$this->protectionValidator = $protectionValidator;
-		$this->permissionExaminer = $permissionExaminer;
+	public function __construct( \MediaWiki\Permissions\PermissionManager $permissionManager = null ) {
+		$this->permissionManager = $permissionManager;
 	}
 
 	/**
-	 * @since 3.1
+	 * @since 3.2
 	 *
-	 * @return []
-	 */
-	public function getErrors() {
-		return $this->errors;
-	}
-
-	/**
-	 * @since 2.5
-	 *
-	 * @param Title $title
-	 * @param User $user
 	 * @param string $action
-	 * @param array &$errors
+	 * @param User|null $user
+	 * @param Title $title
 	 *
-	 * @return boolean
+	 * @return bool
 	 */
-	public function checkQuickPermission( Title $title, User $user, $action, &$errors ) {
+	public function userCan( string $action, User $user = null, Title $title ) : bool {
 
-		$hasUserPermission = $this->hasUserPermission( $title, $user, $action );
-
-		foreach ( $this->errors as $error ) {
-			$errors[] = $error;
+		// @see Title::userCan
+		if ( !$user instanceof User ) {
+			$user = $GLOBALS['wgUser'];
 		}
 
-		return $hasUserPermission;
+		if ( $this->permissionManager !== null ) {
+			return $this->permissionManager->userCan( $action, $user, $title );
+		}
+
+		return $title->userCan( $action, $user );
 	}
 
 	/**
-	 * @since 2.4
+	 * @since 3.2
 	 *
-	 * @param Title $title
 	 * @param User $user
 	 * @param string $action
 	 *
-	 * @return boolean
+	 * @return bool
 	 */
-	public function hasUserPermission( Title $title, User $user, $action ) {
+	public function userHasRight( User $user, string $action = '' ) : bool {
 
-		$this->errors = [];
-
-		if ( $title->getNamespace() === SMW_NS_SCHEMA ) {
-			return $this->checkSchemaNamespacePermission( $title, $user, $action );
+		if ( $this->permissionManager !== null && method_exists( $this->permissionManager, 'userHasRight' ) ) {
+			return $this->permissionManager->userHasRight( $user, $action );
 		}
 
-		$actions = [ 'edit', 'delete', 'move', 'upload' ];
-
-		if ( !in_array( $action, $actions ) ) {
-			return true;
-		}
-
-		if ( $title->getNamespace() === NS_MEDIAWIKI ) {
-			return $this->checkMwNamespacePatternEditPermission( $title, $user, $action );
-		}
-
-		if ( $this->protectionValidator->getCreateProtectionRight() && $title->getNamespace() === SMW_NS_PROPERTY ) {
-			return $this->checkPropertyNamespaceCreatePermission( $title, $user, $action );
-		}
-
-		if ( $title->getNamespace() === NS_CATEGORY ) {
-			return $this->checkChangePropagationProtection( $title, $user, $action );
-		}
-
-		if ( !$title->exists() ) {
-			return true;
-		}
-
-		if ( $title->getNamespace() === SMW_NS_PROPERTY ) {
-			return $this->checkPropertyNamespaceEditPermission( $title, $user, $action );
-		}
-
-		if ( $this->protectionValidator->hasEditProtectionOnNamespace( $title ) ) {
-			return $this->checkEditPermission( $title, $user, $action );
-		}
-
-		return true;
-	}
-
-	private function checkMwNamespacePatternEditPermission( Title $title, User $user, $action ) {
-
-		// @see https://www.semantic-mediawiki.org/wiki/Help:Special_property_Allows_pattern
-		if (
-			$title->getDBKey() !== AllowsPatternValue::REFERENCE_PAGE_ID ||
-			$this->permissionExaminer->userHasRight( $user, 'smw-patternedit' ) ) {
-			return true;
-		}
-
-		$this->errors[] = [ 'smw-patternedit-protection', 'smw-patternedit' ];
-
-		return false;
-	}
-
-	private function checkSchemaNamespacePermission( Title $title, User $user, $action ) {
-
-		if ( !$this->permissionExaminer->userHasRight( $user, 'smw-schemaedit' ) ) {
-			$this->errors[] = [ 'smw-schema-namespace-edit-protection', 'smw-schemaedit' ];
-			return false;
-		}
-
-		// Disallow to change the content model
-		if ( $action === 'editcontentmodel' ) {
-			$this->errors[] = [ 'smw-schema-namespace-editcontentmodel-disallowed' ];
-			return false;
-		}
-
-		if ( $action === 'edit' && $this->protectionValidator->isClassifiedAsImportPerformerProtected( $title, $user ) ) {
-			$this->errors[] = [ 'smw-schema-namespace-edit-protection-by-import-performer' ];
-			return false;
-		}
-
-		return true;
-	}
-
-	private function checkPropertyNamespaceCreatePermission( Title $title, User $user, $action ) {
-
-		$protectionRight = $this->protectionValidator->getCreateProtectionRight();
-
-		if ( $protectionRight === false ) {
-			$protectionRight = $this->protectionValidator->getEditProtectionRight();
-		}
-
-		if ( $this->permissionExaminer->userHasRight( $user, $protectionRight ) ) {
-			return $this->checkPropertyNamespaceEditPermission( $title, $user, $action );
-		}
-
-		$msg = 'smw-create-protection';
-
-		if ( $title->exists() ) {
-			$msg = 'smw-create-protection-exists';
-		}
-
-		$this->errors[] = [ $msg, $title->getText(), $protectionRight ];
-
-		return false;
-	}
-
-	private function checkPropertyNamespaceEditPermission( Title $title, User $user, $action ) {
-
-		// This renders full protection until the ChangePropagationDispatchJob was run
-		if ( !$this->protectionValidator->hasChangePropagationProtection( $title ) ) {
-			return $this->checkEditPermission( $title, $user, $action );
-		}
-
-		$this->errors[] = [ 'smw-change-propagation-protection' ];
-
-		return false;
-	}
-
-	private function checkChangePropagationProtection( Title $title, User $user, $action ) {
-
-		// This renders full protection until the ChangePropagationDispatchJob was run
-		if ( !$this->protectionValidator->hasChangePropagationProtection( $title ) ) {
-			return true;
-		}
-
-		$this->errors[] = [ 'smw-change-propagation-protection' ];
-
-		return false;
-	}
-
-	private function checkEditPermission( Title $title, User $user, $action ) {
-
-		$editProtectionRight = $this->protectionValidator->getEditProtectionRight();
-
-		// @see https://www.semantic-mediawiki.org/wiki/Help:Special_property_Is_edit_protected
-		if (
-			!$this->protectionValidator->hasProtection( $title ) ||
-			$this->permissionExaminer->userHasRight( $user, $editProtectionRight ) ) {
-			return true;
-		}
-
-		$this->errors[] = [ 'smw-edit-protection', $editProtectionRight ];
-
-		return false;
+		return $user->isAllowed( $action );
 	}
 
 }
